app/requirements.txt
typer
python-dotenv
rich
fastapi
loguru
typer
uvicorn
scipy
websockets

huggingface-hub
Pillow
timm
ultralytics
opencv-python~=4.10.0
easyocr
deskew
httpx
soundfile